Python/Python3/requirements.txt
pyserial>=3.4
numpy>=1.22.0
